+++ s3-other.tf
@@ -29,6 +29,26 @@ resource "aws_s3_bucket" "apiary_inventory_bucket" {
           }
        }
     },
+%{if length(var.s3_inventory_customer_accounts) > 0}
+    {
+        "Sid": "S3 inventory customer account permissions",
+        "Effect": "Allow",
+        "Principal": {
+          "AWS": [ "${join("\",\"", formatlist("arn:aws:iam::%s:root", var.s3_inventory_customer_accounts))}" ]
+        },
+        "Action": [
+            "s3:GetBucketLocation",
+            "s3:GetObject",
+            "s3:GetObjectAcl",
+            "s3:GetBucketAcl",
+            "s3:ListBucket"
+        ],
+        "Resource": [
+            "arn:aws:s3:::${local.s3_inventory_bucket}",
+            "arn:aws:s3:::${local.s3_inventory_bucket}/*"
+        ]
+    },
+%{endif}
     {
       "Sid": "DenyUnSecureCommunications",
       "Effect": "Deny",