+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.8.1] - 2021-06-17
+### Added
+- Add support for cross account access to s3 inventory.
+
 ## [6.8.0] - 2021-05-10
 ### Added
 - Add support for Apiary-specific RDS parameter groups.